pgm_graph_inference/requirements.txt
numpy
scipy
torch
torchvision
tqdm
networkx>=2.3
python-igraph